+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+## 0.15.0 (8.12.2014)
+
+- **new feature**: binding dynamic variable `ring.swagger.json-schema/*ignore-missing-mappings*`
+  to true will cause unknown json-schema mappings to be ignored
+- fixed #42
+- updated dependencies:
+```clojure
+[metosin/ring-http-response "0.5.2"] is available but we use "0.5.1"
+[prismatic/schema "0.3.3"] is available but we use "0.3.2"
+[ring/ring-core "1.3.2"] is available but we use "1.3.1"
+```
+
+## 0.14.1 (11.10.2014)
+
+- support for `s/Named`
+- use `[org.tobereplaced/lettercase "1.0.0"]` in favour of `[camel-snake-kebab "0.2.5"]`
+- update deps:
+```clojure
+[prismatic/schema "0.3.2"] is available but we use "0.3.1"
+```
+
+## 0.14.0 (29.10.2014)
+
+- support for `java.lang.Number`, `java.lang.Integer`, `s/Num`
+- Updated deps:
+```clojure
+[slingshot "0.12.1"] is available but we use "0.12.0"
+[metosin/ring-http-response "0.5.1"] is available but we use "0.5.0"
+[prismatic/plumbing "0.3.5"] is available but we use "0.3.4"
+[camel-snake-kebab "0.2.5"] is available but we use "0.2.4"
+[potemkin "0.3.11"] is available but we use "0.3.10"
+[lein-ring "0.8.13"] is available but we use "0.8.11"
+[prismatic/schema "0.3.1"] is available but we use "0.2.6"
+```
+
 ## 0.13.0 (4.9.2014)
 
 - Updated dependencies